requirements.txt
torch
torchdiffeq
numpy
pyyaml
pytest
matplotlib
torch_scatter